requirements.txt
asgiref==3.5.2
Django==4.1.3
django-crispy-forms==1.14.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
dj-database-url==1.0.0
psycopg2==2.9.5
gunicorn==20.1.0
pip==22.3.1
djangorestframework~=3.14.0